finance-reconciliation-agent/requirements.txt
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.15.0
cryptography==41.0.7
python-dotenv==1.0.0
aiohttp==3.9.1
pydantic==2.5.2
prometheus-client==0.19.0
python-json-logger==2.0.7
jinja2==3.1.2
matplotlib==3.8.2
seaborn==0.13.0
requests==2.31.0
